requirements.txt
fastapi
starlette
pydantic
playwright
beautifulsoup4
psycopg2-binary
python-dotenv
requests

